add BaseTest and make other test classes extend BaseTest
