+++ src/test/java/com/solvd/NewInSubMenuTests.java
@@ -1,29 +1,14 @@
 package com.solvd;
 
+import com.solvd.base.BaseTest;
 import com.solvd.components.NewInSubMenu;
 import com.solvd.pages.HomePage;
-import com.solvd.propertiesReader.ConfigReader;
 import com.zebrunner.carina.core.registrar.ownership.MethodOwner;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.Assert;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
 
-import java.net.MalformedURLException;
-import java.net.URL;
-
-public class NewInSubMenuTests {
-    private WebDriver driver;
-
-    @BeforeMethod
-    public void driverSetup() throws MalformedURLException {
-        ChromeOptions options = new ChromeOptions();
-        driver = new RemoteWebDriver(new URL(ConfigReader.getHubUrl()), options);
-    }
+public class NewInSubMenuTests extends BaseTest {
 
     @Test(description = "Display submenu of 'New In' menu category")
     @MethodOwner(owner = "jjanickamatraszek")
@@ -42,10 +27,5 @@ public class NewInSubMenuTests {
         soft.assertEquals(newInSubMenu.getNumberOfSubcategoriesWithoutImg(), 0, "Some subcategories don't have an image");
         soft.assertEquals(newInSubMenu.getNumberOfSubcategoriesWithoutTitle(), 0, "Some subcategories don't have a title");
         soft.assertAll();
-    }
-
-    @AfterMethod
-    public void quitDriver() {
-        driver.quit();
     }
 }
