+++ src/test/java/com/solvd/base/BaseTest.java
@@ -1,0 +1,30 @@
+package com.solvd.base;
+
+import com.solvd.propertiesReader.ConfigReader;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.remote.RemoteWebDriver;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.time.Duration;
+
+public abstract class BaseTest {
+    protected WebDriver driver;
+
+    @BeforeMethod
+    public void driverSetup() throws MalformedURLException {
+        ChromeOptions options = new ChromeOptions();
+        driver = new RemoteWebDriver(new URL(ConfigReader.getHubUrl()), options);
+
+        driver.manage().window().maximize();
+        driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(10));
+    }
+
+    @AfterMethod
+    public void quitDriver() {
+        driver.quit();
+    }
+}
